@@ -5,14 +5,13 @@
 // ignore: unused_import
 import 'dart:ui';
 
-import 'package:cloud_firestore_web/cloud_firestore_web.dart';
 import 'package:firebase_core_web/firebase_core_web.dart';
-
+import 'package:cloud_firestore_web/cloud_firestore_web.dart';
+// ignore: depend_on_referenced_packages
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 
-// ignore: public_member_api_docs
-void registerPlugins(PluginRegistry registry) {
-  FirebaseFirestoreWeb.registerWith(registry.registrarFor(FirebaseFirestoreWeb));
-  FirebaseCoreWeb.registerWith(registry.registrarFor(FirebaseCoreWeb));
-  registry.registerMessageHandler();
+void registerPlugins(Registrar registrar) {
+  FirebaseFirestoreWeb.registerWith(registrar);
+  FirebaseCoreWeb.registerWith(registrar);
+  registrar.registerMessageHandler();
 }
